Create README.md
This app is made for the citizens improve vaccination in india. Aadhar card are scanned to link the newly born babies with their parents and storing all the data related to vaccination on online database and showing it on their app and a reminder will be given of the pending vaccines.